+++ apps/web/src/app/schools/schools.module.ts
@@ -6,10 +6,22 @@ import { SchoolsService } from './schools.service';
 import { SharedModule } from '../shared/shared.module';
 import { NewSchoolComponent } from './new-school/new-school.component';
 import { ListingSchoolsComponent } from './listing-schools/listing-schools.component';
+import { ReactiveFormsModule } from '@angular/forms';
+import { CreateSchoolFormComponent } from './create-school-form/create-school-form.component';
 
 @NgModule({
-  declarations: [SchoolsComponent, NewSchoolComponent, ListingSchoolsComponent],
-  imports: [CommonModule, SchoolsRoutingModule, SharedModule],
+  declarations: [
+    SchoolsComponent,
+    NewSchoolComponent,
+    ListingSchoolsComponent,
+    CreateSchoolFormComponent,
+  ],
+  imports: [
+    CommonModule,
+    SchoolsRoutingModule,
+    SharedModule,
+    ReactiveFormsModule,
+  ],
   providers: [SchoolsService],
 })
 export class SchoolsModule {}